+++ cocos/2d/platform/CCSAXParser.cpp
@@ -114,7 +114,7 @@ bool SAXParser::parse(const char* pXMLData, size_t uDataLength)
 bool SAXParser::parse(const char *pszFile)
 {
     bool ret = false;
-    long size = 0;
+    ssize_t size = 0;
     char* pBuffer = (char*)FileUtils::getInstance()->getFileData(pszFile, "rt", &size);
     if (pBuffer != NULL && size > 0)
     {
